// src/types/callback-query-method.ts
export enum CallbackQueryMethod {
  // TonConnect methods
  CHOOSE_WALLET = 'chose-wallet',
  SELECT_WALLET = 'select-wallet',
  UNIVERSAL_QR = 'universal-qr',

  ANSWER = 'answer',
  FINISH = 'finish',
  SKIP = 'skip',
  START = 'start',
  START_DIALOG = 'start-dialog',
  START_TEST = 'start-test',
  START_LESSON = 'start-lesson',
}